move VariantsSequenceViewer component inside allele table
