+++ src/components/variant/VariantJBrowseLink.js
