+++ src/components/variant/index.js
@@ -1,0 +1,1 @@
+export { default as VariantJBrowseLink } from './VariantJBrowseLink';
